pyproject.toml
[project]
name = "langchian-quiz-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu>=1.12.0",
    "ipykernel>=6.30.1",
    "jupyter>=1.1.1",
    "jupyterlab>=4.4.5",
    "langchain>=1.0.8",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=2.1.9",
    "langgraph>=1.0.3",
    "pymupdf>=1.26.3",
    "python-dotenv>=1.1.1",
    "streamlit>=1.48.0",
]
